feat(EDCO-5): updated stories
